# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Airline Maintenance Timeline API",
    "version": "1.0.0",
    "description": "A microservice for managing airline maintenance timeline data including flights and work packages",
    "contact": {
      "name": "API Support",
      "email": "support@airline-maintenance.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.airline-maintenance.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Health",
      "description": "Service health and monitoring endpoints"
    },
    {
      "name": "Flights",
      "description": "Flight data management"
    },
    {
      "name": "Work Packages",
      "description": "Work package data management"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Check service health status",
        "description": "Returns the health status of the service including database connectivity",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                },
                "examples": {
                  "healthy": {
                    "summary": "Healthy service",
                    "value": {
                      "status": "healthy",
                      "timestamp": "2025-08-12T10:00:00Z",
                      "uptime": 3600.5,
                      "database": "connected"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service is unhealthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                },
                "examples": {
                  "unhealthy": {
                    "summary": "Unhealthy service",
                    "value": {
                      "status": "unhealthy",
                      "timestamp": "2025-08-12T10:00:00Z",
                      "uptime": 3600.5,
                      "database": "disconnected",
                      "error": "Database connection failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/flights": {
      "get": {
        "summary": "Retrieve a list of flights",
        "description": "Get all flights with optional filtering by registration and limiting results",
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "registration",
            "in": "query",
            "description": "Filter by aircraft registration",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9-]{3,10}$",
              "example": "OH-LWA"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of flights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                },
                "examples": {
                  "multiple_flights": {
                    "summary": "Multiple flights response",
                    "value": [
                      {
                        "flightId": 1,
                        "flightNum": "AY101",
                        "registration": "OH-LWA",
                        "schedDepStation": "HEL",
                        "schedArrStation": "LHR",
                        "schedDepTime": "2025-08-08T10:00:00Z",
                        "schedArrTime": "2025-08-08T13:00:00Z"
                      }
                    ]
                  },
                  "empty_result": {
                    "summary": "No flights found",
                    "value": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "invalid_limit": {
                    "summary": "Invalid limit parameter",
                    "value": {
                      "error": "Bad Request",
                      "message": "Limit parameter must be a positive number"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-packages": {
      "get": {
        "summary": "Retrieve a list of work packages",
        "description": "Get all work packages with optional filtering by registration, status and limiting results",
        "tags": [
          "Work Packages"
        ],
        "parameters": [
          {
            "name": "registration",
            "in": "query",
            "description": "Filter by aircraft registration",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9-]{3,10}$",
              "example": "OH-LWA"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by work package status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "In Progress",
                "Completed",
                "Cancelled"
              ],
              "example": "In Progress"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of work packages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkPackage"
                  }
                },
                "examples": {
                  "multiple_packages": {
                    "summary": "Multiple work packages response",
                    "value": [
                      {
                        "workPackageId": 1,
                        "name": "WP-001",
                        "registration": "OH-LWA",
                        "startDateTime": "2025-08-08T13:30:00Z",
                        "endDateTime": "2025-08-08T15:30:00Z",
                        "workOrders": 5,
                        "status": "In Progress"
                      }
                    ]
                  },
                  "empty_result": {
                    "summary": "No work packages found",
                    "value": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Flight": {
        "description": "API Response Types for Airline Maintenance Timeline Service",
        "type": "object",
        "properties": {
          "flightId": {
            "description": "Unique identifier for the flight",
            "type": "number",
            "title": "flightId"
          },
          "flightNum": {
            "description": "Flight number (e.g., AY101)",
            "type": "string",
            "title": "flightNum"
          },
          "registration": {
            "description": "Aircraft registration (e.g., OH-LWA)",
            "type": "string",
            "title": "registration"
          },
          "schedDepStation": {
            "description": "Scheduled departure station IATA code (e.g., HEL)",
            "type": "string",
            "title": "schedDepStation"
          },
          "schedArrStation": {
            "description": "Scheduled arrival station IATA code (e.g., LHR)",
            "type": "string",
            "title": "schedArrStation"
          },
          "schedDepTime": {
            "description": "Scheduled departure time in ISO 8601 format",
            "type": "string",
            "title": "schedDepTime"
          },
          "schedArrTime": {
            "description": "Scheduled arrival time in ISO 8601 format",
            "type": "string",
            "title": "schedArrTime"
          }
        },
        "additionalProperties": false,
        "required": [
          "flightId",
          "flightNum",
          "registration",
          "schedArrStation",
          "schedArrTime",
          "schedDepStation",
          "schedDepTime"
        ],
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "WorkPackage": {
        "type": "object",
        "properties": {
          "workPackageId": {
            "description": "Unique identifier for the work package",
            "type": "number",
            "title": "workPackageId"
          },
          "name": {
            "description": "Work package name (e.g., WP-001)",
            "type": "string",
            "title": "name"
          },
          "registration": {
            "description": "Aircraft registration (e.g., OH-LWA)",
            "type": "string",
            "title": "registration"
          },
          "startDateTime": {
            "description": "Work package start time in ISO 8601 format",
            "type": "string",
            "title": "startDateTime"
          },
          "endDateTime": {
            "description": "Work package end time in ISO 8601 format",
            "type": "string",
            "title": "endDateTime"
          },
          "workOrders": {
            "description": "Number of work orders in the package",
            "type": "number",
            "title": "workOrders"
          },
          "status": {
            "description": "Current status of the work package",
            "enum": [
              "Cancelled",
              "Completed",
              "In Progress",
              "OPEN"
            ],
            "type": "string",
            "title": "status"
          }
        },
        "additionalProperties": false,
        "required": [
          "endDateTime",
          "name",
          "registration",
          "startDateTime",
          "status",
          "workOrders",
          "workPackageId"
        ],
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "HealthStatus": {
        "type": "object",
        "properties": {
          "status": {
            "description": "Overall service health status",
            "enum": [
              "healthy",
              "unhealthy"
            ],
            "type": "string",
            "title": "status"
          },
          "timestamp": {
            "description": "Health check timestamp in ISO 8601 format",
            "type": "string",
            "title": "timestamp"
          },
          "uptime": {
            "description": "Service uptime in seconds",
            "type": "number",
            "title": "uptime"
          },
          "database": {
            "description": "Database connection status",
            "enum": [
              "connected",
              "disconnected"
            ],
            "type": "string",
            "title": "database"
          },
          "error": {
            "description": "Error message if unhealthy",
            "type": "string",
            "title": "error"
          }
        },
        "additionalProperties": false,
        "required": [
          "database",
          "status",
          "timestamp",
          "uptime"
        ],
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "description": "Error type",
            "type": "string",
            "title": "error"
          },
          "message": {
            "description": "Detailed error message",
            "type": "string",
            "title": "message"
          }
        },
        "additionalProperties": false,
        "required": [
          "error",
          "message"
        ],
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "path": {
            "description": "API path that was not found",
            "type": "string",
            "title": "path"
          },
          "method": {
            "description": "HTTP method used",
            "type": "string",
            "title": "method"
          },
          "error": {
            "description": "Error type",
            "type": "string",
            "title": "error"
          },
          "message": {
            "description": "Detailed error message",
            "type": "string",
            "title": "message"
          }
        },
        "additionalProperties": false,
        "required": [
          "error",
          "message",
          "method",
          "path"
        ],
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "FlightQueryParams": {
        "description": "Query Parameters",
        "type": "object",
        "properties": {
          "registration": {
            "description": "Filter by aircraft registration",
            "type": "string",
            "title": "registration"
          },
          "limit": {
            "description": "Maximum number of results to return",
            "type": "number",
            "title": "limit"
          }
        },
        "additionalProperties": false,
        "$schema": "http://json-schema.org/draft-07/schema#"
      },
      "WorkPackageQueryParams": {
        "type": "object",
        "properties": {
          "registration": {
            "description": "Filter by aircraft registration",
            "type": "string",
            "title": "registration"
          },
          "status": {
            "description": "Filter by work package status",
            "enum": [
              "Cancelled",
              "Completed",
              "In Progress",
              "OPEN"
            ],
            "type": "string",
            "title": "status"
          },
          "limit": {
            "description": "Maximum number of results to return",
            "type": "number",
            "title": "limit"
          }
        },
        "additionalProperties": false,
        "$schema": "http://json-schema.org/draft-07/schema#"
      }
    }
  }
}